# lib/api-zod/src/generated/types/hedgeFund.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Stock Research Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface HedgeFund {
  cik: string;
  name: string;
  slug: string;
}