# lib/api-zod/src/generated/types/hedgeFundsResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Stock Research Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { HedgeFund } from "./hedgeFund";

export interface HedgeFundsResponse {
  funds: HedgeFund[];
}