# lib/api-zod/src/generated/types/thirteenFHoldingRow.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Stock Research Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface ThirteenFHoldingRow {
  /** Issuer name from 13F (deduplicated) */
  name: string;
  /** Resolved ticker symbol (null if CUSIP lookup failed) */
  ticker?: string | null;
  cusip: string;
  /** Market value in dollars for the current quarter */
  currentMarketValue: number;
  /** Shares held in the current quarter */
  currentShares: number;
  /** Percentage of total portfolio (current quarter) */
  currentPctAllocation: number;
  /** Market value in dollars for the prior quarter (null if new position) */
  priorMarketValue?: number | null;
  /** Shares held in the prior quarter (null if new position) */
  priorShares?: number | null;
  /** Percentage of total portfolio (prior quarter) */
  priorPctAllocation?: number | null;
  /** Percentage change in shares vs prior quarter (null if new position) */
  pctChangeShares?: number | null;
  /** Color coding: new | increase-high | increase-low | decrease-high | decrease-low | empty string */
  colorClass: string;
}