# lib/api-zod/src/generated/types/thirteenFHoldingsResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Stock Research Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { ThirteenFHoldingRow } from "./thirteenFHoldingRow";

export interface ThirteenFHoldingsResponse {
  fundName: string;
  cik: string;
  currentQ?: string | null;
  priorQ?: string | null;
  /** Total portfolio value in dollars for the current quarter */
  currentTotalValue: number;
  /** Total portfolio value in dollars for the prior quarter */
  priorTotalValue?: number | null;
  /** True if the initial data fetch from SEC EDGAR has not completed yet */
  seedingInProgress: boolean;
  holdings: ThirteenFHoldingRow[];
}