# lib/api-zod/src/generated/types/thirteenFQuartersResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Stock Research Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface ThirteenFQuartersResponse {
  cik: string;
  /** Quarter labels newest first (e.g. ["Q1 2026", "Q4 2025", ...]) */
  quarters: string[];
}